requirements.txt
torch==2.3.0   # Add explicit PyTorch version
markdown>=3.5.2  # 新增Markdown依赖
marker-pdf==1.6.2  
python-pptx>=0.6.21
moviepy==1.0.3
pydub>=0.25.1
PyYAML>=6.0.1
python-dotenv>=1.0.0
googlesearch-python>=1.2.3 # 知识增强
pyttsx3>=2.90        # 文字转语音
openai>=1.12.0  # 添加OpenAI官方SDK
streamlit>=1.28.0
